[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/6879 Former-commit-id: f2337ffd952562664bb7b48848aa609b7e365c17

+++ icu4c/source/test/intltest/normconf.cpp
@@ -170,7 +170,7 @@ void NormalizerConformanceTest::TestConformance(void) {
         }
 
         fields[0]=fields[1]=fields[2]=fields[3]=fields[4].setTo(c);
-        sprintf(lineBuf, "not mentioned code point U+%04lx", c);
+        sprintf(lineBuf, "not mentioned code point U+%04lx", (long)c);
 
         if (checkConformance(fields, UnicodeString(lineBuf, ""))) {
             ++passCount;
